WIP Don't failfast on error; nil checks avoid segfault
The spec i'm working with has a lot of issues and it seemed it may be
simpler to get a list of all errors and fix the spec once rather than
constant iteration. Instead of calling contract.Failf, it collects the
errors for later handling.

Because of spec issues I was getting segfaults on things like missing
request or response bodies, so this adds some functions to help pull out
the request/response bodies and report errors if not present rather than
allowing a segfault panic.